pyproject.toml
[project]
name = "revis"
version = "0.1.0"
description = "Multi-agent research coordination CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "daytona>=0.14.0",
  "pyyaml>=6.0.3",
  "rich>=14.0.0",
  "textual>=3.0.0",
  "tomli-w>=1.2.0",
  "typer>=0.16.0",
]

[project.scripts]
revis = "revis.cli.main:app"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/revis"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
